[Feat] Append git config's email to username
A new variable author_and_email is available. It displays user
information in this format :

git.config.user.name - <git.config.user.email>

To use it simply insert {{author_and_email}} in your templates.

last_modified_by variable is set to that value as a default.

Note : both user.name and user.email should be set in order to make it
work which is considered as a git best practice.